Partial edit Checklist Name
