+++ src/App.jsx
@@ -1,56 +1,12 @@
-import { useEffect, useState } from "react";
 import Layout from "./components/Layout";
-import AddTask from "./components/AddTask";
-import Tasks from "./components/Tasks";
-import { v4 } from "uuid";
+import Checklist from "./components/Checklist";
 
 function App() {
-  const [tasks, setTasks] = useState(
-    JSON.parse(localStorage.getItem("tasks")) || []
-  );
-  useEffect(() => {
-    localStorage.setItem("tasks", JSON.stringify(tasks));
-  }, [tasks]);
-
-  function onTaskClick(taskId) {
-    const updatedTasks = tasks.map((task) => {
-      if (task.id === taskId) {
-        return { ...task, isCompleted: !task.isCompleted };
-      }
-      return task;
-    });
-
-    setTasks(updatedTasks);
-  }
-
-  function onTaskDelete(taskId) {
-    const updatedTasks = tasks.filter((task) => task.id !== taskId);
-    setTasks(updatedTasks);
-  }
-
-  function onAddItemClick(title, description) {
-    if (!title.trim()) return;
-
-    const newTask = {
-      id: v4(),
-      title,
-      description,
-      isCompleted: false,
-    };
-
-    setTasks([...tasks, newTask]);
-  }
-
   return (
     <>
       <Layout>
         <section className="w-[500px] space-y-4">
-          <AddTask onAddItemClick={onAddItemClick} />
-          <Tasks
-            tasks={tasks}
-            onTaskClick={onTaskClick}
-            onTaskDelete={onTaskDelete}
-          />
+          <Checklist />
         </section>
       </Layout>
     </>
