+++ src/components/AddTask.jsx
@@ -6,9 +6,6 @@ function AddTask({ onAddItemClick }) {
 
   return (
     <div className="flex flex-col p-6 space-y-4 bg-bg-card rounded-md text-xl">
-      <h3 className="text-2xl text-center font-bold text-tx-title ">
-        Your Checklist
-      </h3>
       <input
         type="text"
         value={title}
